table views, routing and exception changes
Overwrote a layer on top of the RouteAwere mixin to make it more generic. This was done by overwriting a layer on "ModalRoute.of", but this is a temporary solution. It is suitable for simple use until an official solution is provided by Flutter. Added state management to TablesView. Added a generic exception class.